testsuite-fast: provider name empty fixed + testsuite exit value fixed

+++ tests/run_tests.sh
@@ -28,21 +28,45 @@ source run_requirements.sh
 #    cd ..
 #done
 
+WHITE='\033[0m'
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+
 while read curr; do
-    #echo "$curr"
     cd "$curr"
-    #ls
     source ./run.sh
+
+    RET=$?
+    if [ "$RET" -eq "1" ]
+    then
+        cd ..
+        rm .out*
+        kubectl delete -f required-test-cloud-credential.yaml
+        kubectl delete -f secret_cloud.yaml
+        kubectl delete -f required-test-kubernetes-profile.yaml
+        kubectl delete -f required-test-user.yaml
+        kubectl delete -f required-test-organization.yaml
+        rm required*
+        rm secret*
+
+        printf "\n${RED}TESTSUITE FAILED !${WHITE}\n\n"
+        exit 1
+    fi
+
     cd ..
 done < ".out_test"
 
 rm .out*
-rm secret*
 
 kubectl delete -f required-test-cloud-credential.yaml
-
-rm required-test-cloud-credential.yaml
+kubectl delete -f secret_cloud.yaml
 
 kubectl delete -f required-test-kubernetes-profile.yaml
 kubectl delete -f required-test-user.yaml
 kubectl delete -f required-test-organization.yaml
+
+rm required*
+rm secret*
+
+printf "\n${GREEN}TESTSUITE PASSED SUCCESSFULLY !!!${WHITE}\n\n"
+exit 0
